Change the relationship between backup and container names
Backups will correspond to a container called '<container name>-backup-container'

Backups will fail to be created or updated if there already exists a container with the
same name in the storage account or if it is not possible to connect to check.

On name editing, everything from the old container will be copied into the new container.
This is not implemented yet