backup-password: Set passwords on back-end as opposed to front
Now we inject the correct backup password when the backup is being added or edited

There's also a modification on progress steps when restoring from e-Notariado. Now
the first step corresponds to retrieving the list of remote backups, when that returns,
we go to the second step which is to select the backup we want to restore.